requirements.txt
opencv-python
loguru
pyautogui
torch
torchvision
pillow
PyQt5

